(#91958) Process: include string resources properly in tests
